// ui/src/lib/booking-hotel-api/model/room-type-save-request.ts
/**
 * main-app API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RoomTypeSaveRequest { 
    id?: number;
    name: string;
    hotelId: number;
    bedTypeId: number;
    count: number;
}