// ui/src/lib/booking-hotel-api/model/room-type-model.ts
/**
 * main-app API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {FacilityModel1} from './facility-model1';
import {BedTypeModel1} from './bed-type-model1';


export interface RoomTypeModel {
    id: number;
    hotelId: number;
    bedTypeId: number;
    name: string;
    count: number;
    bedType?: BedTypeModel1;
    facilities?: Array<FacilityModel1>;
    createdAt: string;
    updatedAt: string;
    removed: boolean;
}